pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyfibot"
version = "0.1.0"
description = "A Python IRC bot with modular plugin system"
requires-python = ">=3.8"
dependencies = [
    "twisted>=24.0.0",
    "requests",
    "pyyaml",
    "jsonschema>=1.0.0",
    "beautifulsoup4",
    "pyopenssl",
    "ndg-httpsclient",
    "pyasn1",
    "tvdb-api",
    "repoze.lru",
    "pygeoip",
    "sqlalchemy",
    "dataset",
    "python-instagram",
    "simplejson>=3.0.0",
    "httplib2>=0.8.0",
    "feedparser>=5.1.3",
    "service-identity",
    "python-dateutil>=2.2",
    "pytz",
    "pytvmaze",
    "gitpython>=3.1.0",
    "requests-cache>=1.0.0",
    "tenacity>=8.0.0",
    "lxml>=4.6.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "flake8",
    "pylint",
    "pep8",
    "autopep8",
    "pytest",
    "pytest-cov",
    "coverage",
    "vcrpy",
]

[tool.hatch.build.targets.sdist]
include = [
    "/pyfibot",
    "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["pyfibot"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = false
ignore_missing_imports = true
ignore_errors = false
strict_optional = false
files = ["pyfibot/botcore.py", "pyfibot/pyfibot.py"]
exclude = ["pyfibot/modules/"]

[[tool.mypy.overrides]]
module = "pyfibot.modules.*"
ignore_errors = true


[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20241230",
    "types-requests>=2.31.0.6",
]
